tests/wpt/tests/navigation-api/navigate-event/intercept-after-dispatch.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  navigation.onnavigate = t.step_func(e => {
    t.step_timeout(t.step_func_done(() => {
      assert_equals(e.eventPhase, Event.NONE);
      assert_throws_dom("InvalidStateError", () => e.intercept());
    }), 0);
  });

  location.href = "#1";
}, "event.intercept() throws if used after the dispatch phase");
</script>
